// auth/src/shared-resources/common/enum/global-configuration-name.enum.ts
export enum GlobalConfigurationNameEnum {
  ACCESS_ATTEMPT_LIMIT = 'access-attempt-limit',
  ACCESS_ATTEMPT_LOCKOUT_DURATION = 'access-attempt-lockout-duration',
  ACCESS_ATTEMPT_DURATION = 'access-attempt-duration',
  ALLOW_MISS_SCAN_TIME = 'allow-miss-scan-time',
  ALLOW_DAY_OFF_DAYS = 'allow-day-off-days',
  ALLOW_LATE_SCAN_IN = 'allow-late-scan-in',
  ALLOW_BREAK_TIME_SCAN_AFTER_CHECK_IN = 'allow-break-time-scan-after-check-in',
  ALLOW_BEFORE_AND_AFTER_START_SCAN_DURATION = 'allow-before-and-after-start-scan-duration',
  ALLOW_BEFORE_AND_AFTER_END_SCAN_DURATION = 'allow-before-and-after-end-scan-duration',
  ALLOW_CHECK_OUT_EARLY = 'allow-check-out-early',
  ENABLE_TELEGRAM_NOTIFICATION = 'enable-telegram-notification',
  ENABLE_STRONG_PASSWORD = 'enable-strong-password',
  TELEGRAM_TOKEN = 'telegram-token',
  TELEGRAM_CHAT_ID = 'telegram-chat-id',
  EMAIL_NOTIFICATION = 'email-notification',
  MAIL_HOST = 'mail-host',
  MAIL_PORT = 'mail-port',
  MAIL_USER = 'mail-user',
  MAIL_PASS = 'mail-password',
  // FIREBASE
  FIREBASE_TYPE = 'firebase_type',
  FIREBASE_PROJECT_ID = 'firebase_project_id',
  FIREBASE_PRIVATE_KEY_ID = 'firebase_private_key_id',
  FIREBASE_PRIVATE_KEY = 'firebase_private_key',
  FIREBASE_CLIENT_EMAIL = 'firebase_client_email',
  FIREBASE_CLIENT_ID = 'firebase_client_id',
  FIREBASE_JWT_EXPIRE = 'firebase_jwt_expire',
  REAL_TIME_NOTIFICATION = 'real_time_notification',
  // Mekong
  SMS_PASSWORD = 'sms-password',
  SMS_SERVER_URL = 'sms-server-url',
  SMS_USERNAME = 'sms-username',
  SMS_SENDER = 'sms-sender',
  SMS_IN = 'sms-in',
  SMS_CUSTOM_DATA = 'custom-data',
  OVERTIME_REQUEST_DURATION = 'overtime-request-duration',
  REMOVE_DUPLICATION_SCAN_TIME_IN_MINUTE = 'remove-duplication-scan-time-in-minute'
}